test(nfr): give the two refusal suites one fixture instead of two
The COMPAT.1 revision matrix needs the same gateway state the SEC.1 suite
builds -- a 45-line AppState literal naming twenty-odd fields, plus the request
helpers around it. Copying that into a second integration test would have put
two literals of that shape in the tree, and the next field the struct gains
lands in one of them.

tests/common/mod.rs now holds it, moved rather than rewritten: the same text,
with pub added to the items and to the import list so a suite needs one
`use common::*;` and cannot drift into a second list kept in step by hand.
Cargo compiles the module once per test binary, so items one suite does not use
are dead in that binary by construction -- allow(dead_code) carries that reason,
rather than suppressing something worth fixing.

NOT COMPILED. The disk guard still halts cargo (4.2 GB free, below the 5 GB
floor). This is a move with no behaviour of its own, which is the only reason it
is defensible to land unverified; every assertion in the suite is unchanged.
cargo test is owed on both binaries before any passing claim.
